@@ -2,4 +2,4 @@ package com.example.smart_tailor
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity : FlutterActivity()
